[12128] Fix more warnings - close #4
* Remove warning: enumeration value '<value>' not handled in switch
* Remove warning: enumeral and non-enumeral type in conditional expression
* Remove warning: too many arguments for format


Conflicts:
	src/game/Channel.cpp
	src/game/GameObject.cpp
	src/game/Pet.cpp
	src/game/Player.cpp
	src/game/QuestHandler.cpp
	src/game/Spell.cpp
	src/game/SpellAuras.cpp

+++ src/game/Channel.cpp
@@ -579,7 +579,7 @@ void Channel::Say(ObjectGuid p, const char *what, uint32 lang, bool ignoreOn)
     data << p;
     data << uint32(messageLength);
     data << what;
-    data << uint8(plr ? plr->GetChatTag() : CHAT_TAG_NONE);
+    data << uint8(plr ? plr->GetChatTag() : uint8(CHAT_TAG_NONE));
 
     SendToAll(&data, IsOn(p) && !m_players[p].IsModerator() ? p : ObjectGuid());
 }
